Support Microsoft Visual C.
Do not include stdint.h on Windows, make our own compatibility macros
for stdint types and inline. Convert BLAKE2 implementations to,
basically, C89.

+++ impl/blake2.h
@@ -14,8 +14,8 @@
 #ifndef __BLAKE2_H__
 #define __BLAKE2_H__
 
+#include "blake2-config.h"
 #include <stddef.h>
-#include <stdint.h>
 
 #if defined(_MSC_VER)
 #define ALIGN(x) __declspec(align(x))
@@ -143,7 +143,7 @@ extern "C" {
   int blake2sp( uint8_t *out, const void *in, const void *key, const uint8_t outlen, const uint64_t inlen, uint8_t keylen );
   int blake2bp( uint8_t *out, const void *in, const void *key, const uint8_t outlen, const uint64_t inlen, uint8_t keylen );
 
-  static inline int blake2( uint8_t *out, const void *in, const void *key, const uint8_t outlen, const uint64_t inlen, uint8_t keylen )
+  static inline int blake2( uint8_t *out, const void *in, const void *key, const uint8_t outlen, const size_t inlen, uint8_t keylen )
   {
     return blake2b( out, in, key, outlen, inlen, keylen );
   }
